This commit contains too many changes, sorry
It ensures that all imported storage are imported directly from
models.__init__.py and not just from storage. This ensured that
amenities that were associated with places using append() were also
committed because previously these were added to the table without
committing and thus weren't visible on the table. The commit also makes
changes to the relationships between the tables but these are
inconsequential and were just made in an attempt to debug the above said
issue. close() methods were added to dbstorage and filestorage. If the
storage type isn't db there is a getter method called cities that was
added to State class and it returns a list of City objects from storage
that are linked to the current state. The rest of the changes are mostly
modifications to the relationships which I've explained above.